Release 8.240 of Pythia.

+++ src/SusyLesHouches.cc
@@ -155,7 +155,7 @@ int SusyLesHouches::readFile(istream& is, int verboseIn,
         string pdgString = blockIn.substr(pdgBegin,pdgEnd-pdgBegin);
         istringstream linestream(pdgString);
         // Create and add new block with this code as zero'th entry
-        LHblock<int> newQnumbers;
+        LHblock<double> newQnumbers;
         newQnumbers.set(0,linestream);
         qnumbers.push_back(newQnumbers);
         // Default name: PDG code
